Trait dslab_faas::cpu::CpuPolicy

pub trait CpuPolicy {
    // Required methods
    fn init(&self, cores: u32) -> Box<dyn CpuPolicy>;
    fn get_load(&self) -> f64;
    fn on_new_invocation(
        &mut self,
        invocation: &mut Invocation,
        container: &mut Container,
        time: f64,
        ctx: &mut SimulationContext
    );
    fn on_invocation_end(
        &mut self,
        invocation: &mut Invocation,
        container: &mut Container,
        time: f64,
        ctx: &mut SimulationContext
    );
}

CpuPolicy governs CPU sharing, computes invocation progress and manages invocation end events.

Required Methods

fn init(&self, cores: u32) -> Box<dyn CpuPolicy>

Returns a clean initialized instance of the policy.

fn get_load(&self) -> f64

Returns the current CPU load.

fn on_new_invocation( &mut self, invocation: &mut Invocation, container: &mut Container, time: f64, ctx: &mut SimulationContext )

This method is called whenever there is a new invocation to run on this CPU.

fn on_invocation_end( &mut self, invocation: &mut Invocation, container: &mut Container, time: f64, ctx: &mut SimulationContext )

This method is called whenever an invocation on this CPU stops executing.