Module dslab_faas::extra

Various implementations of DSLab FaaS traits, including trace formats, schedulers and coldstart policies.

Modules

azure_trace_2019

Functions for parsing Azure Functions 2019 trace and generating experiments using it.

azure_trace_2021

Functions for parsing Azure Functions 2021 trace and generating experiments using it.

hermod

Scheduler implementation from “Hermod: principled and practical scheduling for serverless functions” paper.

hybrid_histogram

Coldstart policy implementation from “Serverless in the Wild: Characterizing and Optimizing the Serverless Workload at a Large Cloud Provider” paper.

opendc_trace

OpenDC FaaS trace format parser.

resolvers

Additional resolvers for YAML configs.

synthetic_trace

Synthetic trace generator.