Struct dslab_faas::extra::azure_trace_2019::Azure2019TraceConfig

pub struct Azure2019TraceConfig {
    pub time_period: u64,
    pub time_skip: u64,
    pub duration_generator: DurationGenerator,
    pub start_generator: StartGenerator,
    pub app_preferences: Vec<AppPreference>,
    pub concurrency_level: usize,
    pub random_seed: u64,
    pub memory_name: String,
    pub force_fixed_memory: Option<u64>,
    pub cold_start_latency: f64,
    pub rps: Option<f64>,
}

Struct with Azure 2019 trace settings.

Fields
time_period: u64

Simulation time period in minutes (only integer numbers are supported).

time_skip: u64

This option allows skipping an integer number of minutes from the start of the trace. It may be useful if you want to split trace into small 1-hour experiments for example.

duration_generator: DurationGenerator

This option controls the method used to generate execution durations.

start_generator: StartGenerator

This option controls the method used to generate start times.

app_preferences: Vec<AppPreference>

This option controls which apps to use for trace generation.

concurrency_level: usize

This option sets concurrency level for all apps in the trace.

random_seed: u64

This option sets the seed used to initialize random generator.

memory_name: String

This option sets name for the memory resource.

force_fixed_memory: Option<u64>

This option forces trace generator to use given amount of memory for all apps.

cold_start_latency: f64

Cold start latency, currently it’s the same for all apps.

rps: Option<f64>

If rps is not None, trace generator attempts to scale trace to the given number of requests per second by either removing random requests or duplicating random requests (yes, that doesn’t sound very solid).