Module dslab_faas::invoker

A component that manages incoming invocations on a host.

Structs

DequeuedInvocation

Previously-queued invocation that is finally able to be executed.

FIFOInvoker

FIFOInvoker repeatedly tries to invoke the oldest queued invocation.

NaiveInvoker

NaiveInvoker iterates over all queued invocations and tries to invoke each of them. In case of large queues it may be very slow, use FIFOInvoker instead.

Enums

InvokerDecision

Invoker’s decision on new function invocation.

Traits

Invoker

Invoker handles invocations at host level. It chooses containers for execution, deploys new containers and manages invocation queue.

Functions

default_invoker_resolver

Creates Invoker from a string containing its name and parameters.