Module dslab_faas::parallel

Utilities for running multiple experiments in parallel.

Structs

ParallelConfig

Similar to crate::config::Config, but ensures that all simulation components implement Send trait.

ParallelHostConfig

Similar to crate::config::HostConfig, but only accepts invokers with Send trait.

Functions

parallel_simulation

Runs parallel simulations in a thread pool with a separate worker for each config.

parallel_simulation_n_workers

Runs parallel simulations in a thread pool with n_workers worker threads.

parallel_simulation_raw

Similar to parallel_simulation, but for raw configs.

parallel_simulation_raw_n_workers

Similar to parallel_simulation_n_workers, but for raw configs.

parallel_simulation_yaml

Similar to parallel_simulation, but for YAML configs.

parallel_simulation_yaml_n_workers

Similar to parallel_simulation_n_workers, but for YAML configs.