Module dslab_faas::resource

Resource model.

The simulator considers all resources to be renewable and all resources except CPU to be strictly isolated among containers.

Structs

Resource

A simple strictly isolated renewable resource.

ResourceConsumer

A model of a container or other entity that consumes resources offered by the provider.

ResourceNameResolver

Transforms resource name to resource id in current simulation.

ResourceProvider

A model of a host or other entity that provides several resources to the consumers.

ResourceRequirement

Requirement for allocating specified amount of a resource.