Module dslab_faas::scheduler

Scheduler trait and several simple implementations.

Structs

ApplicationHasher

Hashes application ids, used to map applications to hosts.

BasicScheduler

Chooses the first invoker that can hotstart the invocation, otherwise chooses the first invoker that can deploy the container.

LeastLoadedScheduler

Chooses a host with the least CPU usage.

LocalityBasedScheduler

Picks a host based on the application hash. In case host number i can’t invoke, the scheduler considers host number (i + step) % hosts.len().

RandomScheduler

Picks a host uniformly at random.

RoundRobinScheduler

Chooses hosts in a circular fashion.

Traits

Scheduler

Chooses a host to run new invocation of some function from given application.

Functions

default_scheduler_resolver

Creates Scheduler from a string containing its name and parameters.