Struct dslab_faas::simulation::ServerlessSimulation

pub struct ServerlessSimulation { /* private fields */ }

Main FaaS simulation class.

Implementations

impl ServerlessSimulation

pub fn new(sim: Simulation, config: Config) -> Self

Creates new simulation.

pub fn try_resolve_resource_name(&self, name: &str) -> Option<usize>

Resolves resource name into id if it exists.

pub fn create_resource(&mut self, name: &str, available: u64) -> Resource

Creates new resource.

pub fn create_resource_requirement( &mut self, name: &str, needed: u64 ) -> ResourceRequirement

Creates new resource requirement.

pub fn get_invocation(&self, id: usize) -> Invocation

Returns invocation by its id.

pub fn get_invocations(&self, id: Range<usize>) -> Vec<Invocation>

Returns all invocations with ids in given range.

pub fn stats(&self) -> Stats

Returns simulation metrics.

pub fn global_stats(&self) -> GlobalStats

Returns global metrics.

pub fn invocation_stats(&self) -> InvocationStats

Returns global invocation metrics.

pub fn add_host( &mut self, invoker: Option<Box<dyn Invoker>>, resources: ResourceProvider, cores: u32 )

Adds a new crate::host::Host.

pub fn add_function(&mut self, f: Function) -> usize

Adds a new crate::function::Function.

pub fn add_app_with_single_function(&mut self, app: Application) -> usize

Adds a new crate::function::Application with a single function.

pub fn add_app(&mut self, app: Application) -> usize

Adds a new crate::function::Application.

pub fn load_trace(&mut self, trace: &dyn Trace) -> Range<usize>

Returns a consecutive range of indices of new invocations. Invocations have consecutive ids in increasing order of time.

pub fn send_requests_from_ordered_iter( &mut self, iterator: &mut dyn Iterator<Item = RequestData> ) -> Range<usize>

This function provides a way to send invocation requests to a special ordered deque inside the simulation, which works faster than the default heap. Returns a consecutive range of indices of new invocations, invocations have consecutive ids that follow their order inside the iterator.

pub fn send_invocation_request( &mut self, id: usize, duration: f64, time: f64 ) -> usize

Sends a single invocation request.

pub fn set_simulation_end(&mut self, time: f64)

Simulation end event is useful in case you have a no-unloading policy and you want metrics like wasted resource time to be correct at the end of simulation (of course, you have to provide correct time).

pub fn step(&mut self) -> bool

Makes one simulation step.

pub fn steps(&mut self, step_count: u64) -> bool

Makes given number of simulation steps.

pub fn step_for_duration(&mut self, duration: f64)

Makes simulation steps for given time duration.

pub fn step_until_no_events(&mut self)

Makes simulation steps until no events remain.

pub fn event_count(&self) -> u64

Returns number of events in the simulation.