Struct dslab_faas::container::ContainerManager

pub struct ContainerManager { /* private fields */ }

Manages containers on a single host.

Implementations

impl ContainerManager

pub fn new( host_id: usize, resources: ResourceProvider, ctx: Rc<RefCell<SimulationContext>> ) -> Self

Creates new ContainerManager on a specified host.

pub fn can_allocate(&self, resources: &ResourceConsumer) -> bool

Checks whether it is possible to create a new container with given resources.

pub fn get_total_resource(&self, id: usize) -> u64

Returns total amount of a resource.

pub fn dec_active_invocations(&mut self)

Decrements active invocations count.

pub fn inc_active_invocations(&mut self)

Increments active invocations count.

pub fn active_invocation_count(&self) -> usize

Returns active invocations count.

pub fn get_container(&self, id: usize) -> Option<&Container>

Returns a reference to a container specified by id if it exists.

pub fn get_container_mut(&mut self, id: usize) -> Option<&mut Container>

Returns a mutable reference to a container specified by id if it exists.

pub fn get_containers(&mut self) -> &mut FxIndexMap<usize, Container>

Returns a mutable reference to the map that stores all existing containers.

pub fn get_possible_containers( &self, app: &Application, allow_deploying: bool ) -> PossibleContainerIterator<'_>

Returns an iterator over running containers that can accommodate one more invocation of given app. If allow_deploying is true, also returns containers that are being deployed.

pub fn try_deploy( &mut self, app: &Application, time: f64 ) -> Option<(usize, f64)>

Tries to deploy a new container for given app.

pub fn reserve_container(&mut self, id: usize, request: usize)

Reserves a deploying container for a new invocation.

pub fn count_reservations(&self, id: usize) -> usize

Counts reserved invocations for a deploying container.

pub fn take_reservations(&mut self, id: usize) -> Option<Vec<usize>>

Returns and removes reserved invocations for a deploying container.

pub fn delete_container(&mut self, id: usize)

Deletes a container.

pub fn try_move_container_to_free(&mut self, id: usize)

Moves a container to free list if it was full.

pub fn move_container_to_full(&mut self, id: usize)

Moves a container to full list.