Module dslab_faas::cpu

CPU sharing models.

We use something similar to CPU shares in cgroups, but instead of shares we operate with (shares/core_shares), i.e. if the container has 512 shares and each core amounts to 1024 shares, the share of the container equals 0.5. If the container allows concurrent invocations, each invocation gets an equal part of the container share. The exact CPU sharing model depends on the used CpuPolicy.

Structs

ContendedCpuPolicy

CPU shares of active containers may exceed the number of cores, in this case the invocations are slowed down. We assume that CPU sharing is fair: each invocation makes progress according to its share.

Cpu

Just a wrapper over CpuPolicy.

IgnoredCpuPolicy

This policy ignores contention for CPU resources. All invocations work as if executed on some abstract infinite-core CPU without any contention.

IsolatedCpuPolicy

CPU shares of active containers should not exceed the number of cores. If this limit is exceeded, extra invocations are queued until some cores are freed. This model currently ignores possible effects related to multiple concurrent invocations on the same container.

Traits

CpuPolicy

CpuPolicy governs CPU sharing, computes invocation progress and manages invocation end events.

Functions

default_cpu_policy_resolver

Creates CpuPolicy from a string containing its name and parameters.